integrated code with the actual images and captions data

+++ include/part2/embedding/matrix.py
@@ -122,6 +122,6 @@ class SimilarityMatrix(Matrix):
 
     def __init__(self, pairs, nr_images, nr_captions):
         super().__init__()
-        self.matrix = np.zeros(shape=(nr_images, nr_captions))
+        self.matrix = np.zeros(shape=(nr_images, nr_captions), dtype=np.int8)
         for (index_image, index_caption) in pairs:
             self.matrix[index_image, index_caption] = 1  # value is 1 if pair exists AKA image and caption are similar
